// src/types/omlet.type.ts
export enum OmletType {
    CONFIGURATION = 'configuration',
    CONFIGURATION_GENERAL = 'configuration_general',
    CONFIGURATION_CONNECTIVITY = 'configuration_connectivity',
    CONFIGURATION_DOOR = 'configuration_door',
    CONFIGURATION_LIGHT = 'configuration_light',
    STATE = 'state',
    STATE_GENERAL = 'state_general',
    STATE_CONNECTIVITY = 'state_connectivity',
    STATE_DOOR = 'state_door',
    STATE_LIGHT = 'state_light',
    DEVICE = 'device',
    ACTION = 'action',
    GROUP_USER = 'group_user',
    GROUP_SUBSET = 'group_subset',
    GROUP = 'group',
    NOTIFICATION_PERIOD = 'notification_period',
    NOTIFICATION_SETTINGS = 'notification_settings',
    USER = 'user',
}