// src/types/configuration.feeder.ts
export interface ConfigurationFeeder {
    mode: string;
    openLightLevel: number;
    closeLightLevel: number;
    openTime1: string;
    openTime2: string;
    openTime3: string;
    openTime4: string;
    closeTime1: string;
    closeTime2: string;
    closeTime3: string;
    closeTime4: string;
    colour: string;
}