// src/types/configuration.fan.ts
export interface ConfigurationFan {
    mode: string;
    manualSpeed: number;
    timeOn1: string;
    timeOn2: string;
    timeOn3: string;
    timeOn4: string;
    timeOff1: string;
    timeOff2: string;
    timeOff3: string;
    timeOff4: string;
    timeSpeed1: number;
    timeSpeed2: number;
    timeSpeed3: number;
    timeSpeed4: number;
    tempOn: number;
    tempOff: number;
    tempSpeed: number;
}